+++ src/server/components/dashboard/listsItems/controllers.ts
@@ -143,73 +143,6 @@ export async function listItemPostController(req: Request, res: Response) {
   return res.redirect(`${listItemUrl}/confirm`);
 }
 
-export async function listItemPinController(req: Request, res: Response) {
-  const { action } = req.body;
-  const userId = req.user!.userData.id;
-  const listItem = res.locals.listItem!;
-  const isPinned = action === "pin";
-  const { listItemUrl, listIndexUrl } = res.locals;
-
-  try {
-    await handlePinListItem(listItem.id, userId, isPinned);
-
-    req.flash(
-      "successBannerTitle",
-      `${listItem.jsonData.organisationName} has been ${isPinned ? "pinned" : "unpinned"}`
-    );
-    req.flash("successBannerHeading", `${isPinned ? "Pinned" : "Unpinned"}`);
-    req.flash("successBannerColour", "blue");
-    return res.redirect(listIndexUrl);
-  } catch (error) {
-    logger.error(`listItemPinController: ${userId} failed to ${action} user, ${error}`);
-    req.flash("errorMsg", `${listItem.jsonData.organisationName} could not be updated. ${(error as Error).message}`);
-    return res.redirect(listItemUrl);
-  }
-}
-
-export async function handlePinListItem(id: number, userId: User["id"], isPinned: boolean): Promise<ListItem> {
-  if (userId === undefined) {
-    throw new Error("deleteListItem Error: userId is undefined");
-  }
-
-  try {
-    const auditEvent = recordListItemEvent(
-      {
-        eventName: isPinned ? "pin" : "unpin",
-        itemId: id,
-        userId,
-      },
-      isPinned ? AuditEvent.PINNED : AuditEvent.UNPINNED
-    );
-
-    const connectOrDisconnect = isPinned ? "connect" : "disconnect";
-    const pinOrUnpinEvent = isPinned ? EVENTS.PINNED(userId) : EVENTS.PINNED(userId);
-
-    const [listItem] = await prisma.$transaction([
-      prisma.listItem.update({
-        where: {
-          id,
-        },
-        data: {
-          pinnedBy: {
-            [connectOrDisconnect]: [{ id: userId }],
-          },
-          history: {
-            create: [pinOrUnpinEvent],
-          },
-        },
-      }),
-      auditEvent,
-    ]);
-
-    return listItem;
-  } catch (e: any) {
-    logger.error(`deleteListItem Error ${e.message}`);
-
-    throw new Error(`Failed to ${isPinned ? "pin" : "unpinned"} item`);
-  }
-}
-
 export async function listItemUpdateController(req: Request, res: Response): Promise<void> {
   const userId = req.user!.id;
   const listItem = res.locals.listItem;
